requirements.txt
streamlit
openai
gtts
requests
pillow
ffmpeg-python
